requirements.txt
gitpython
pyyaml
pydantic